+++ changelog.md
@@ -1,1 +1,1 @@
-- Incorrect model state being used on neoforge.
+- Support for `tint` in models, can either be a tint instead of an object of r,g,b,a for a static color